Email-Delivery-Agent-main/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
google-generativeai==0.3.2
pydantic==2.5.0
pydantic-settings==2.1.0
sqlalchemy==2.0.31
python-multipart==0.0.6
aiosqlite==0.19.0
